requirements.txt
streamlit
python-dotenv
flask
Flask[async]